implemented policy for contact model (#45)
* implemented policy for contact model

* implemented policies for all models

* updated redirections to dash"

+++ app/Livewire/Companies/Show.php
@@ -24,12 +24,16 @@ class Show extends Component
 
     public function mount(Company $company)
     {
+        $this->authorize('view', $company);
+        
         $this->company = $company;
         $this->title = $company->name;
     }
 
     public function removeFromCompany($contactId)
     {
+        $this->authorize('update', $this->company);
+        
         $contact = Contact::find($contactId);
         $contact->company()->dissociate();
         $contact->save();
@@ -38,6 +42,8 @@ class Show extends Component
 
     public function delete()
     {
+        $this->authorize('delete', $this->company);
+        
         $this->company->delete();
 
         session()->flash('success', 'Company deleted successfully');
